+++ tests/test_aws_sns_sqs_service_with_credentials_with_custom_envelope.py
@@ -12,11 +12,11 @@ from run_test_service_helper import start_service
     not os.environ.get("TOMODACHI_TEST_AWS_ACCESS_KEY_ID") or not os.environ.get("TOMODACHI_TEST_AWS_ACCESS_SECRET"),
     reason="AWS configuration options missing in environment",
 )
-def test_start_aws_sns_sqs_service_with_credentials_with_custom_protocol(
+def test_start_aws_sns_sqs_service_with_credentials_with_custom_envelope(
     monkeypatch: Any, capsys: Any, loop: Any
 ) -> None:
     services, future = start_service(
-        "tests/services/aws_sns_sqs_service_with_credentials_with_custom_protocol.py", monkeypatch
+        "tests/services/aws_sns_sqs_service_with_credentials_with_custom_envelope.py", monkeypatch
     )
 
     assert services is not None